+++ src/importer/xml/mod.rs
@@ -28,6 +28,8 @@ pub struct ImportXML {
     named_orderings: BTreeSet<String>,
     #[serde(default)]
     skip_names: BTreeSet<String>,
+    #[serde(default)]
+    use_ids: bool,
 }
 
 const MODULE_NAME: &str = "import_xml";
@@ -70,6 +72,7 @@ impl ImportXML {
         let mut ordering_ends_at: BTreeMap<String, String> = BTreeMap::new();
         let mut node_count = 1;
         let mut char_buffer = String::new();
+        let mut ignore_chars = false;
         // parsing
         let f = File::open(path)?;
         let mut parser_cfg = ParserConfig::new();
@@ -87,22 +90,24 @@ impl ImportXML {
                 } => {
                     let lookup = name.to_string();
                     if self.skip_names.contains(&lookup) {
+                        ignore_chars = true;
                         continue;
                     }
-                    let parent_node_id = match node_stack.last() {
-                        None => {
-                            return Err(AnnattoError::Import {
-                                reason: format!(
-                                    "{} cannot be mapped. Node stack is empty.",
-                                    path.to_string_lossy()
-                                ),
-                                importer: MODULE_NAME.to_string(),
-                                path: path.to_path_buf(),
-                            })
+                    ignore_chars = false;
+                    let generic_id = format!("{}#n{node_count}", doc_node_id.to_string());
+                    let node_id = if self.use_ids {
+                        if let Some(attr) = attributes
+                            .iter()
+                            .filter(|a| a.name.to_string().as_str() == "id")
+                            .last()
+                        {
+                            format!("{}#{}", doc_node_id.to_string(), attr.value.to_string())
+                        } else {
+                            generic_id
                         }
-                        Some(v) => v.to_string(),
+                    } else {
+                        generic_id
                     };
-                    let node_id = format!("{}#n{node_count}", doc_node_id.to_string());
                     update.add_event(UpdateEvent::AddNode {
                         node_name: node_id.to_string(),
                         node_type: "node".to_string(),
@@ -117,38 +122,56 @@ impl ImportXML {
                             anno_value: value.trim().to_string(),
                         })?;
                     }
-                    update.add_event(UpdateEvent::AddEdge {
-                        source_node: node_id.to_string(),
-                        target_node: parent_node_id.to_string(),
-                        layer: ANNIS_NS.to_string(),
-                        component_type: AnnotationComponentType::PartOf.to_string(),
-                        component_name: "".to_string(),
-                    })?;
-                    if node_stack.len() > 1 {
+                    if self.default_ordering != lookup {
+                        update.add_event(UpdateEvent::AddNodeLabel {
+                            node_name: node_id.to_string(),
+                            anno_ns: ANNIS_NS.to_string(),
+                            anno_name: "layer".to_string(),
+                            anno_value: GENERIC_NS.to_string(),
+                        })?;
+                    } else if node_stack.len() > 1 {
+                        let empty_node = (&node_id).replace("#", "#t_");
+                        update.add_event(UpdateEvent::AddNode {
+                            node_name: empty_node.to_string(),
+                            node_type: "node".to_string(),
+                        })?;
                         update.add_event(UpdateEvent::AddEdge {
-                            source_node: parent_node_id.to_string(),
-                            target_node: node_id.to_string(),
-                            layer: DEFAULT_NS.to_string(),
+                            source_node: node_id.to_string(),
+                            target_node: empty_node.to_string(),
+                            layer: ANNIS_NS.to_string(),
                             component_type: AnnotationComponentType::Coverage.to_string(),
                             component_name: "".to_string(),
                         })?;
-                    }
-                    if self.default_ordering == lookup {
+                        for ascendent_node in &node_stack[1..] {
+                            update.add_event(UpdateEvent::AddEdge {
+                                source_node: ascendent_node.to_string(),
+                                target_node: empty_node.to_string(),
+                                layer: DEFAULT_NS.to_string(),
+                                component_type: AnnotationComponentType::Coverage.to_string(),
+                                component_name: "".to_string(),
+                            })?;
+                        }
                         if let Some(source_node_id) = ordering_ends_at.get(&"".to_string()) {
                             update.add_event(UpdateEvent::AddEdge {
                                 source_node: source_node_id.to_string(),
-                                target_node: node_id.to_string(),
+                                target_node: empty_node.to_string(),
                                 layer: ANNIS_NS.to_string(),
                                 component_type: AnnotationComponentType::Ordering.to_string(),
                                 component_name: "".to_string(),
                             })?;
                         }
-                        ordering_ends_at.insert("".to_string(), node_id.to_string());
+                        ordering_ends_at.insert("".to_string(), empty_node.to_string());
+                        update.add_event(UpdateEvent::AddNodeLabel {
+                            node_name: empty_node.to_string(),
+                            anno_ns: ANNIS_NS.to_string(),
+                            anno_name: "layer".to_string(),
+                            anno_value: "default_layer".to_string(),
+                        })?;
                         update.add_event(UpdateEvent::AddNodeLabel {
                             node_name: node_id.to_string(),
                             anno_ns: ANNIS_NS.to_string(),
                             anno_name: "layer".to_string(),
-                            anno_value: "default_layer".to_string(),
+                            anno_value: GENERIC_NS.to_string(),
                         })?;
                     }
                     if self.named_orderings.contains(&lookup) {
@@ -163,6 +186,7 @@ impl ImportXML {
                         }
                         ordering_ends_at.insert(lookup.to_string(), node_id.to_string());
                         update.add_event(UpdateEvent::AddNodeLabel {
+                            // might overwrite a previous layer annotation, but that is okay
                             node_name: node_id.to_string(),
                             anno_ns: ANNIS_NS.to_string(),
                             anno_name: "layer".to_string(),
@@ -173,6 +197,9 @@ impl ImportXML {
                     node_count += 1;
                 }
                 xml::reader::XmlEvent::EndElement { name } => {
+                    if self.skip_names.contains(&name.to_string()) {
+                        continue;
+                    }
                     if let Some(node_id) = node_stack.pop() {
                         let value = char_buffer.to_string();
                         char_buffer.clear();
@@ -183,10 +210,9 @@ impl ImportXML {
                             anno_value: value.to_string(),
                         })?;
                         let lookup = name.to_string();
-                        if self.default_ordering == lookup || self.named_orderings.contains(&lookup)
-                        {
+                        if self.default_ordering == lookup {
                             update.add_event(UpdateEvent::AddNodeLabel {
-                                node_name: node_id.to_string(),
+                                node_name: node_id.replace("#", "#t_"),
                                 anno_ns: ANNIS_NS.to_string(),
                                 anno_name: "tok".to_string(),
                                 anno_value: value,
@@ -194,7 +220,11 @@ impl ImportXML {
                         }
                     }
                 }
-                xml::reader::XmlEvent::Characters(chars) => char_buffer.push_str(chars.as_str()),
+                xml::reader::XmlEvent::Characters(chars) => {
+                    if !ignore_chars {
+                        char_buffer.push_str(chars.as_str());
+                    }
+                }
                 xml::reader::XmlEvent::EndDocument => {
                     progress.worked(1)?;
                     break;