+++ src/importer/xml/tests.rs
@@ -1,4 +1,4 @@
-use std::{collections::BTreeSet, path::Path};
+use std::path::Path;
 
 use insta::assert_snapshot;
 
@@ -9,8 +9,19 @@ fn test_generic_xml() {
     let actual = import_as_graphml_string(
         ImportXML {
             default_ordering: "w".to_string(),
-            named_orderings: vec!["w".to_string()].into_iter().collect::<BTreeSet<String>>(),
-            skip_names: BTreeSet::new()
+            named_orderings: vec!["w".to_string()].into_iter().collect(),
+            skip_names: vec![
+                "document".to_string(),
+                "P".to_string(),
+                "AUTHOR".to_string(),
+                "URL".to_string(),
+                "QUOTE".to_string(),
+                "TITLE".to_string(),
+                "TRANSLATOR".to_string(),
+            ]
+            .into_iter()
+            .collect(),
+            use_ids: true,
         },
         Path::new("tests/data/import/generic_xml/"),
         None,