+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `path` as a import format now triggers the embedding of path names as nodes into the graph; this is supposed to help to represent configuration files for ANNIS
 - import `path` adds an `annis::file` annotation
 - import `path` adds part-of edges
+- very basic implementation of a generic xml importer
+- import opus sentence alignments
+- graph op `enumerate` to enumerate nodes, i. e., add numeric annotations to results of one or multiple queries
 
 ### Fixed